Better Makefile, use deps
