+++ Makefile
@@ -10,21 +10,16 @@ CXXFLAGS			:= -std=c++17 -Wall -Wextra -O3
 
 override SRCS		:=							\
 				main.cpp						\
+				worker.cpp						\
 				utils.cpp						\
 				print.cpp						\
 				args.cpp						\
 				executor.cpp					\
 				arguments_generator.cpp			\
 
-override HEADERS	:=							\
-				complexity.hpp					\
-				executor.hpp					\
-				arguments_generator.hpp			\
-				thread_safe_random.hpp			\
+override OBJS		:= $(addprefix build/, $(SRCS:.cpp=.o))
 
-override HEADERS	:= $(addprefix src/,$(HEADERS))
-
-override OBJS		:= $(addprefix obj/, $(SRCS:.cpp=.o))
+override DEPS		:= $(OBJS:.o=.d)
 
 override OBJDIRS	:= $(sort $(dir $(OBJS)))
 
@@ -32,8 +27,8 @@ override OBJDIRS	:= $(sort $(dir $(OBJS)))
 
 all:		$(NAME)
 
-obj/%.o:	src/%.cpp $(HEADERS)
-			$(CXX) $(CXXFLAGS) -c $< -o $@ -Iincludes
+build/%.o:	src/%.cpp
+			$(CXX) $(CXXFLAGS) -c $< -o $@ -MMD -MP
 
 $(OBJS):	| $(OBJDIRS)
 
@@ -44,7 +39,7 @@ $(NAME):	$(OBJS)
 			$(CXX) $(CXXFLAGS) -o $@ $(OBJS)
 
 clean:
-			rm -rf obj
+			rm -rf build
 
 fclean:		clean
 			rm -f $(NAME)
@@ -52,3 +47,5 @@ fclean:		clean
 re:			fclean all
 
 .PHONY:		all clean fclean re
+
+-include $(DEPS)
