+++ src/arguments_generator.hpp
@@ -8,40 +8,10 @@
 #include <unordered_set>
 #include <memory_resource>
 
-#include "thread_safe_random.hpp"
+#include "arguments_iterator.hpp"
 
 class ArgumentsGenerator;
-
-class ArgumentsIterator
-{
-    friend class ArgumentsGenerator;
-
-private:
-    char* const m_begin;
-    char* const m_end;
-    char* m_pos;
-    size_t m_step;
-
-    ArgumentsIterator(char* begin, char* end, const size_t step) : m_begin(begin), m_end(end), m_pos(begin),
-                                                                   m_step(step)
-    {
-    }
-
-public:
-    [[nodiscard]] auto begin() const -> const char* { return m_begin; }
-    [[nodiscard]] auto end() const -> const char* { return m_end; }
-    auto reset() -> void { m_pos = m_begin; }
-
-    auto next() -> char*
-    {
-        if (m_pos >= m_end)
-            return nullptr;
-
-        char* tmp = m_pos;
-        m_pos += m_step;
-        return tmp;
-    }
-};
+class ThreadSafeRandom;
 
 class ArgumentsGenerator
 {
